refactor(mpsc): organize the bounded channel by endpoint
Split the 650-line bounded module along the channel's own structure:
mod.rs wires up the channel, sender.rs and receiver.rs hold the two
public endpoints, and the two internal mechanisms get one file each —
semaphore.rs for capacity, buffer.rs for storage.

The sender's wait machinery moves into the semaphore as Acquire, the
in-flight counterpart of acquire(), so all sentinel and wait-queue
interlocking lives in one place. Verb-named operations replace
noun-shaped internals: Reservation becomes Acquire, and the receiver's
single-attempt core try_pop becomes pull.
